+++ sycl/include/sycl/queue.hpp
@@ -1196,11 +1196,11 @@ event submitAssertCapture(queue &Self, event &Event, queue *SecondaryQueue,
         const char *Func = AH->Func[0] ? AH->Func : "<unknown func>";
 
         fprintf(stderr,
-                "%s:%d: %s: global id: [%" PRIu64 ",%" PRIu64 ",%" PRIu64
+                "%s:%lld: %s: global id: [%" PRIu64 ",%" PRIu64 ",%" PRIu64
                 "], local id: [%" PRIu64 ",%" PRIu64 ",%" PRIu64 "] "
                 "Assertion `%s` failed.\n",
-                File, AH->Line, Func, AH->GID0, AH->GID1, AH->GID2, AH->LID0,
-                AH->LID1, AH->LID2, Expr);
+                File, (long long int)AH->Line, Func, AH->GID0, AH->GID1,
+                AH->GID2, AH->LID0, AH->LID1, AH->LID2, Expr);
         fflush(stderr);
         abort(); // no need to release memory as it's abort anyway
       }
